# Add indexes on feeder_results and limit_violations tables