--- a/src/main/resources/db/changelog/changesets/changelog_2023-03-31T11:38:24Z.xml
+++ b/src/main/resources/db/changelog/changesets/changelog_2023-03-31T11:38:24Z.xml
@@ -1,0 +1,11 @@
+<?xml version="1.1" encoding="UTF-8" standalone="no"?>
+<databaseChangeLog xmlns="http://www.liquibase.org/xml/ns/dbchangelog" xmlns:ext="http://www.liquibase.org/xml/ns/dbchangelog-ext" xmlns:pro="http://www.liquibase.org/xml/ns/pro" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog-ext http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-ext.xsd http://www.liquibase.org/xml/ns/pro http://www.liquibase.org/xml/ns/pro/liquibase-pro-4.1.xsd http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-4.1.xsd">
+    <changeSet author="bouzolssyl (generated)" id="1680262718004-1">
+        <createIndex indexName="feeder_results_fault_result_idx" tableName="feeder_results">
+            <column name="fault_result_entity_fault_result_uuid"/>
+        </createIndex>
+        <createIndex indexName="limit_violations_fault_result_idx" tableName="limit_violations">
+            <column name="fault_result_entity_fault_result_uuid"/>
+        </createIndex>
+    </changeSet>
+</databaseChangeLog>