--- a/src/main/resources/db/changelog/db.changelog-master.yaml
+++ b/src/main/resources/db/changelog/db.changelog-master.yaml
@@ -6,3 +6,6 @@ databaseChangeLog:
   - include:
       file: changesets/changelog_2023-01-13T13:22:26Z.xml
       relativeToChangelogFile: true
+  - include:
+      file: changesets/changelog_2023-03-31T11:38:24Z.xml
+      relativeToChangelogFile: true