For perf optim when getting results

ex: with a classical network
**Before**
| nb of shortcircuit results| compute time | store time | get results time |
|--------------|-----------|------------|-----------|
| 1 | 15s | 2s | 2s |
| 5 | 15s | 2s | 3s |
| 10 | 15s | 2s | 5s |

**After**
| nb of shortcircuit results| compute time | store time | get results time |
|--------------|-----------|------------|-----------|
| 1 | 15s | 2s | **1s** |
| 5 | 15s | 2s | **1s** |
| 10 | 15s | 2s | **1s** |